Refactor tool_shed.metadata.metadata_generator into Galaxy.
